fixed bugs in dequeue and enqueue functions

+++ circular_array.h
@@ -11,9 +11,9 @@
 typedef struct process{
     unsigned int time_arr;
     char name[MAX_NAME_LEN];
-    unsigned int serv_time;
+    long serv_time;
     int mem_req;
-    unsigned int serv_time_remaining;
+    long serv_time_remaining;
 }process_t;
 
 struct circular_array {
